refactor() : 패키지 구조 변경 - gift 패키지 아래 business, presentation, persistence 패키지로 계층 분리 - 각 패키지 아래 dto, model과 로직 패키지 분리
