+++ src/main/java/gift/business/model/Product.java
@@ -1,4 +1,4 @@
-package gift.domain;
+package gift.business.model;
 
 public class Product {
