+++ src/main/java/gift/persistence/repository/ProductRepository.java
@@ -1,6 +1,6 @@
-package gift.repository;
+package gift.persistence.repository;
 
-import gift.domain.Product;
+import gift.business.model.Product;
 
 import java.util.List;
