+++ src/main/java/gift/presentation/controller/ProductManagerController.java
@@ -1,8 +1,8 @@
-package gift.controller;
+package gift.presentation.controller;
 
 
-import gift.dto.ResponseProductDto;
-import gift.service.ProductService;
+import gift.presentation.dto.ResponseProductDto;
+import gift.business.service.ProductService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
