+++ src/main/java/gift/presentation/dto/RequestProductIdsDto.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.presentation.dto;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Size;
