+++ src/main/java/gift/presentation/restcontroller/ProductController.java
@@ -1,9 +1,9 @@
-package gift.restcontroller;
+package gift.presentation.restcontroller;
 
-import gift.dto.RequestProductDto;
-import gift.dto.RequestProductIdsDto;
-import gift.dto.ResponseProductDto;
-import gift.service.ProductService;
+import gift.presentation.dto.RequestProductDto;
+import gift.presentation.dto.RequestProductIdsDto;
+import gift.presentation.dto.ResponseProductDto;
+import gift.business.service.ProductService;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
